mini-mvp/pyproject.toml
[project]
name = "mini-mvp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "litellm>=1.72.6",
    "python-fasthtml>=0.12.19",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
